Replace dynamic module-attribution block with WP native blocks.
Previously, the custom module-attribution block was used, so that the
module attribution string could be generated on the server when the
post was rendered. However, this resulted in a system where the admin
was unable to modify the content of the block. Instead, we move to
a system where the attribution content inserted during cloning
is made up of regular WP native blocks.

A backward compatibility layer is in place that converts the old
block markup to the new, the first time that the post is loaded.